+++ packages/core/src/index.ts
@@ -104,7 +104,7 @@ export { moduleMetadataIntegration } from './integrations/metadata';
 export { requestDataIntegration } from './integrations/requestdata';
 export { metrics } from './metrics/exports';
 export type { MetricData } from './metrics/exports';
-export { metricsDefault } from './metrics/exports-node';
+export { metricsDefault } from './metrics/exports-default';
 export { BrowserMetricsAggregator } from './metrics/browser-aggregator';
 
 /** @deprecated Import the integration function directly, e.g. `inboundFiltersIntegration()` instead of `new Integrations.InboundFilter(). */
